pyproject.toml
[tool.poetry]
name = "finnub-test"
version = "0.1.0"
description = ""
authors = ["Pasquale <p3dimaria@hotmail.it>"]

[tool.poetry.dependencies]
python = "^3.8"
plotext = "^1.0.10"
finnhub-python = "^2.2.1"
sqlalchemy = "^1.3.19"
configargparse = "^1.2.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
